timlib: Test had bug, module count was a long, should be int

+++ tim/test/Main.c
@@ -109,7 +109,7 @@ char *text; {
 int main(int argc,char *argv[]) {
 
 int i, j, found_device;
-unsigned long cnt, mod, cbl, dbg, ind, enb, jtg, oby;
+unsigned int cnt, mod, cbl, dbg, ind, enb, jtg, oby;
 char *cp, *ep;
 char host[49];
 char tmpb[CMD_BUF_SIZE];
